Refactors writing of files inside Coordinator
