+++ phase1-coordinator/src/storage/disk.rs
@@ -1,6 +1,6 @@
 use crate::{
     environment::Environment,
-    objects::{ContributionFileSignature, Round},
+    objects::{ContributionFileSignature, Round, ContributionInfo, TrimmedContributionInfo},
     storage::{
         ContributionLocator, ContributionSignatureLocator, Locator, Object, ObjectReader, ObjectWriter, StorageLocator,
         StorageObject,
@@ -198,6 +198,14 @@ impl Disk {
 
                 let contribution_file_signature: ContributionFileSignature = serde_json::from_slice(&file_bytes)?;
                 Ok(Object::ContributionFileSignature(contribution_file_signature))
+            },
+            Locator::ContributionInfoFile { round_height: _ } => {
+                let contribution_info: ContributionInfo = serde_json::from_slice(&file_bytes)?;
+                Ok(Object::ContributionInfoFile(contribution_info))
+            },
+            Locator::ContributionsInfoSummary => {
+                let summary: Vec<TrimmedContributionInfo> = serde_json::from_slice(&file_bytes)?;
+                Ok(Object::ContributionsInfoSummary(summary))
             }
         };
 
@@ -633,6 +641,8 @@ impl StorageLocator for DiskResolver {
                     ),
                 }
             }
+            Locator::ContributionInfoFile { round_height } => format!("{}/contributors/namada_contributor_info_round_{}.json", self.base, round_height),
+            Locator::ContributionsInfoSummary => format!("{}/contributors.json", self.base)
         };
         // Sanitize the path.
         LocatorPath::try_from(Path::new(&path))
